refactor: mv basis test as separate file
